+++ gdal-sys/src/lib.rs
@@ -2,6 +2,4 @@
 #![allow(non_camel_case_types)]
 #![allow(non_snake_case)]
 
-extern crate libc;
-
 include!(concat!(env!("OUT_DIR"), "/bindings.rs"));