+++ src/vector/gdal_to_geo.rs
@@ -1,6 +1,6 @@
+use crate::vector::Geometry;
 use gdal_sys::{self, OGRwkbGeometryType};
 use geo_types;
-use vector::Geometry;
 
 impl From<Geometry> for geo_types::Geometry<f64> {
     fn from(geo: Geometry) -> geo_types::Geometry<f64> {